Introduce PeepAnimation compilation unit
