+++ src/openrct2/scripting/bindings/entity/ScGuest.cpp
@@ -15,6 +15,7 @@
     #include "../../../entity/Guest.h"
     #include "../../../localisation/Formatting.h"
     #include "../../../peep/PeepAnimationData.h"
+    #include "../../../peep/PeepAnimations.h"
     #include "../../../ride/RideEntry.h"
 
 namespace OpenRCT2::Scripting
@@ -145,35 +146,6 @@ namespace OpenRCT2::Scripting
         { "nice_ride_deprecated", PeepThoughtType::NiceRideDeprecated },
         { "excited_deprecated", PeepThoughtType::ExcitedDeprecated },
         { "here_we_are", PeepThoughtType::HereWeAre },
-    });
-
-    static const DukEnumMap<PeepAnimationType> availableGuestAnimations({
-        { "walking", PeepAnimationType::Walking },
-        { "checkTime", PeepAnimationType::CheckTime },
-        { "watchRide", PeepAnimationType::WatchRide },
-        { "eatFood", PeepAnimationType::EatFood },
-        { "shakeHead", PeepAnimationType::ShakeHead },
-        { "emptyPockets", PeepAnimationType::EmptyPockets },
-        { "holdMat", PeepAnimationType::HoldMat },
-        { "sittingIdle", PeepAnimationType::SittingIdle },
-        { "sittingEatFood", PeepAnimationType::SittingEatFood },
-        { "sittingLookAroundLeft", PeepAnimationType::SittingLookAroundLeft },
-        { "sittingLookAroundRight", PeepAnimationType::SittingLookAroundRight },
-        { "hanging", PeepAnimationType::Hanging },
-        { "wow", PeepAnimationType::Wow },
-        { "throwUp", PeepAnimationType::ThrowUp },
-        { "jump", PeepAnimationType::Jump },
-        { "drowning", PeepAnimationType::Drowning },
-        { "joy", PeepAnimationType::Joy },
-        { "readMap", PeepAnimationType::ReadMap },
-        { "wave", PeepAnimationType::Wave },
-        { "wave2", PeepAnimationType::Wave2 },
-        { "takePhoto", PeepAnimationType::TakePhoto },
-        { "clap", PeepAnimationType::Clap },
-        { "disgust", PeepAnimationType::Disgust },
-        { "drawPicture", PeepAnimationType::DrawPicture },
-        { "beingWatched", PeepAnimationType::BeingWatched },
-        { "withdrawMoney", PeepAnimationType::WithdrawMoney },
     });
 
     ScGuest::ScGuest(EntityId id)
@@ -874,7 +846,7 @@ namespace OpenRCT2::Scripting
     std::vector<std::string> ScGuest::availableAnimations_get() const
     {
         std::vector<std::string> availableAnimations{};
-        for (auto& animation : availableGuestAnimations)
+        for (auto& animation : getAnimationsByPeepType(AnimationPeepType::Guest))
         {
             availableAnimations.push_back(std::string(animation.first));
         }
@@ -885,6 +857,7 @@ namespace OpenRCT2::Scripting
     {
         std::vector<uint32_t> spriteIds{};
 
+        auto& availableGuestAnimations = getAnimationsByPeepType(AnimationPeepType::Guest);
         auto animationType = availableGuestAnimations.TryGet(groupKey);
         if (animationType == std::nullopt)
         {
@@ -894,7 +867,7 @@ namespace OpenRCT2::Scripting
         auto peep = GetPeep();
         if (peep != nullptr)
         {
-            auto& animationGroup = GetPeepAnimation(peep->AnimationGroup, *animationType);
+            const auto& animationGroup = GetPeepAnimation(peep->AnimationGroup, *animationType);
             for (auto frameOffset : animationGroup.frame_offsets)
             {
                 auto imageId = animationGroup.base_image;
@@ -917,6 +890,7 @@ namespace OpenRCT2::Scripting
             return nullptr;
         }
 
+        auto& availableGuestAnimations = getAnimationsByPeepType(AnimationPeepType::Guest);
         std::string_view action = availableGuestAnimations[peep->AnimationType];
 
         // Special consideration for sitting peeps
@@ -931,6 +905,7 @@ namespace OpenRCT2::Scripting
     {
         ThrowIfGameStateNotMutable();
 
+        auto& availableGuestAnimations = getAnimationsByPeepType(AnimationPeepType::Guest);
         auto newType = availableGuestAnimations.TryGet(groupKey);
         if (newType == std::nullopt)
         {
@@ -946,7 +921,7 @@ namespace OpenRCT2::Scripting
         else
             peep->AnimationFrameNum = offset;
 
-        auto& animationGroup = GetPeepAnimation(peep->AnimationGroup, peep->AnimationType);
+        const auto& animationGroup = GetPeepAnimation(peep->AnimationGroup, peep->AnimationType);
         peep->AnimationImageIdOffset = animationGroup.frame_offsets[offset];
         peep->UpdateSpriteBoundingBox();
     }
@@ -971,7 +946,7 @@ namespace OpenRCT2::Scripting
 
         auto* peep = GetGuest();
 
-        auto& animationGroup = GetPeepAnimation(peep->AnimationGroup, peep->AnimationType);
+        const auto& animationGroup = GetPeepAnimation(peep->AnimationGroup, peep->AnimationType);
         auto length = animationGroup.frame_offsets.size();
         offset %= length;
 
@@ -992,7 +967,7 @@ namespace OpenRCT2::Scripting
             return 0;
         }
 
-        auto& animationGroup = GetPeepAnimation(peep->AnimationGroup, peep->AnimationType);
+        const auto& animationGroup = GetPeepAnimation(peep->AnimationGroup, peep->AnimationType);
         return static_cast<uint8_t>(animationGroup.frame_offsets.size());
     }
